+++ src/controllers/routing/route/dtos/update.dto.ts
@@ -53,6 +53,10 @@ export class RouteControllerUpdateDto {
     @IsOptional()
     style?: string;
 
+    @IsNumber()
+    @IsOptional()
+    firewallApplyToId?: number
+
     @IsArray()
     @IsOptional()
     @ValidateNested({