+++ src/models/routing/route/route.service.ts
@@ -59,6 +59,7 @@ export interface ICreateRoute {
     active?: boolean;
     comment?: string;
     style?: string;
+    firewallApplyToId?: number;
     ipObjIds?: {id: number, order: number}[];
     ipObjGroupIds?: {id: number, order: number}[];
     openVPNIds?: {id: number, order: number}[];
@@ -73,6 +74,7 @@ interface IUpdateRoute {
     gatewayId?: number;
     interfaceId?: number;
     style?: string;
+    firewallApplyToId?: number;
     ipObjIds?: {id: number, order: number}[];
     ipObjGroupIds?: {id: number, order: number}[];
     openVPNIds?: {id: number, order: number}[];
@@ -153,6 +155,7 @@ export class RouteService extends Service {
             ipObjGroupIds: data.ipObjGroupIds,
             openVPNIds: data.openVPNIds,
             openVPNPrefixIds: data.openVPNPrefixIds,
+            firewallApplyToId: data.firewallApplyToId,
             interfaceId: data.interfaceId,
         })
 
@@ -213,6 +216,10 @@ export class RouteService extends Service {
                 order: item.order
             } as RouteToOpenVPNPrefix));
         }
+
+        await this.validateFirewallApplyToId(firewall, data);
+        route.firewallApplyToId = data.firewallApplyToId;
+        
 
         if (Object.prototype.hasOwnProperty.call(data, "interfaceId")) {
             if (data.interfaceId !== null) {
@@ -623,6 +630,26 @@ export class RouteService extends Service {
         }
         
         
+        if (Object.keys(errors).length > 0) {
+            throw new ValidationException('The given data was invalid', errors);
+        }
+    }
+
+    protected async validateFirewallApplyToId(firewall: Firewall, data: IUpdateRoute): Promise<void> {
+        const errors: ErrorBag = {};
+
+        if(!data.firewallApplyToId){
+            return;
+        } 
+
+        const firewallApplyToId: Firewall = await getRepository(Firewall).createQueryBuilder('firewall')
+        .where("firewall.id = :id", { id: data.firewallApplyToId })
+        .getOne()
+
+       if(firewallApplyToId.clusterId !== firewall.clusterId){
+            errors[`firewallApplyToId`] = ['This firewall does not belong to cluster']
+        } 
+
         if (Object.keys(errors).length > 0) {
             throw new ValidationException('The given data was invalid', errors);
         }