+++ CHANGELOG.md
@@ -7,6 +7,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased] -
 
 ### Added
+- Allow option 'apply to' for routes and routing policy.
 - Download and show policy script routes.
 
 ### Changed
@@ -17,6 +18,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Backup retention policy task is scheduled.
 - Fixed tree repair process.
 - Updated easy-rsa package from 3.0.6 to 3.1.0.
+
 
 ## [1.5.1] - 2022-04-28
 ### Fixed